+++ src/libponyc/ast/symtab.c
@@ -352,7 +352,7 @@ static void symbol_serialise_trace(pony_ctx_t* ctx, void* object)
 }
 
 static void symbol_serialise(pony_ctx_t* ctx, void* object, void* buf,
-  size_t offset, int mutability)
+  uint64_t offset, int mutability)
 {
   (void)mutability;
